changed use case registering, discord command registering  and new commands

+++ generic/models/substance.go
@@ -1,11 +1,11 @@
 package models
 
 type Substance struct {
-	ID                       uint 
-	Label				     string
-	Value                    string
-	RecommendedDosageMin     uint 
-	RecommendedDosageMax     uint 
-	RecommendedPauseMin 	 uint 
-	RecommendedPauseMax      uint 
+	ID                   uint
+	Label                string
+	Value                string
+	RecommendedDosageMin float64
+	RecommendedDosageMax float64
+	RecommendedPauseMin  float64
+	RecommendedPauseMax  float64
 }
